requirements.txt
aiogram
fastapi
uvicorn
supabase
python-dotenv
python-multipart